+++ src/main/java/org/folio/config/ApplicationConfig.java
@@ -6,4 +6,5 @@ import org.springframework.context.annotation.Configuration;
 @Configuration
 @ComponentScan(basePackages = {
   "org.folio.rest.impl"})
-public class ApplicationConfig {}
+public class ApplicationConfig {
+}