+++ src/main/java/org/folio/rest.impl/DataExportImpl.java
@@ -2,7 +2,6 @@ package org.folio.rest.impl;
 
 import io.vertx.core.AsyncResult;
 import io.vertx.core.Context;
-import io.vertx.core.Future;
 import io.vertx.core.Handler;
 import io.vertx.core.Vertx;
 import io.vertx.core.json.JsonObject;
@@ -11,8 +10,12 @@ import io.vertx.core.logging.LoggerFactory;
 import org.folio.rest.jaxrs.model.ExportRequest;
 import org.folio.rest.jaxrs.resource.DataExport;
 import org.folio.service.manager.ExportManager;
+import org.folio.util.ExceptionToResponseMapper;
+
 import javax.ws.rs.core.Response;
 import java.util.Map;
+
+import static io.vertx.core.Future.succeededFuture;
 
 public class DataExportImpl implements DataExport {
   private static final Logger LOGGER = LoggerFactory.getLogger(DataExportImpl.class);
@@ -25,11 +28,20 @@ public class DataExportImpl implements DataExport {
 
   @Override
   public void postDataExportExport(ExportRequest entity, Map<String, String> okapiHeaders, Handler<AsyncResult<Response>> asyncResultHandler, Context vertxContext) {
-    LOGGER.info("Starting the data-export process for the request: {}", entity);
-    exportManager.startExport(JsonObject.mapFrom(entity), JsonObject.mapFrom(okapiHeaders));
-    Future.succeededFuture()
-      .map(PostDataExportExportResponse.respond204())
-      .map(Response.class::cast)
-      .setHandler(asyncResultHandler);
+
+    vertxContext.runOnContext(c -> {
+      try {
+        LOGGER.info("Starting the data-export process for the request: {}", entity);
+        exportManager.startExport(JsonObject.mapFrom(entity), JsonObject.mapFrom(okapiHeaders));
+        succeededFuture()
+          .map(PostDataExportExportResponse.respond204())
+          .map(Response.class::cast)
+          .otherwise(ExceptionToResponseMapper::map)
+          .setHandler(asyncResultHandler);
+      } catch (Exception exception) {
+        asyncResultHandler.handle(succeededFuture(
+          ExceptionToResponseMapper.map(exception)));
+      }
+    });
   }
 }