+++ src/main/java/org/folio/service/fileupload/storage/FileStorageService.java
@@ -2,9 +2,19 @@ package org.folio.service.fileupload.storage;
 
 import org.folio.service.fileupload.reader.SourceStreamReader;
 
+/**
+ * File storage service.
+ * Service to save files and reading files being stored.
+ */
 public interface FileStorageService {
 
-  default SourceStreamReader getSourceReader() {
-    return new SourceStreamReader() {};
+  /**
+   * Returns an instance of the source reader.
+   * @return SourceStreamReader
+   * @see SourceStreamReader
+   */
+  default SourceStreamReader getReader() {
+    return new SourceStreamReader() {
+    };
   }
 }