+++ NEWS.md
@@ -1,2 +1,2 @@
 ## 2020-01-27 v0.0.1-SNAPSHOT
-# MDEXP-35 Created the main export flow
+* Created the main export flow